move navigation to ship instead of profile
